--- lib/middleman_extension.rb
@@ -1,6 +1,0 @@
-require "middleman-core"
-
-Middleman::Extensions.register :async_loader do
-  require 'middleman-async-loader.rb'
-  Middleman::AsyncLoader
-end
